use session context handle table provider
